Fix handling of `fail_silently`.
This was a combination of 2 issues.

The NVP response encodes the first error code in `l_errorcode0`, not `l_errorcode`, so the `flag_code` was not being set correctly.

Also the `l_longmessage0` in the response was just "Invalid profile status for cancel action", despite PayPal's own documentation claiming it should be exactly as written (i.e. including "; profile should be active or suspended").

So fix the error code bug so we get `flag_code`, and then easily access that instead of trusting the message.